requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.22
pypdf2>=3.0.1
python-dotenv>=1.0.0
tiktoken>=0.5.2
openai>=1.6.1

